+++ packages/ui/src/components/index.ts
@@ -1,7 +1,7 @@
 export * from "./ui/badges";
 export * from "./ui/button";
-export * from "./ui/category_tab";
-export * from "./ui/count_badge";
+export * from "./ui/category-tab";
+export * from "./ui/count-badge";
 export * from "./ui/empty";
 export * from "./ui/filter";
 export * from "./ui/gnb";
@@ -9,7 +9,7 @@ export * from "./ui/input-field";
 export * from "./ui/input-image";
 export * from "./ui/input-text-area";
 export * from "./ui/pagination";
-export * from "./ui/progress_bar";
+export * from "./ui/progress-bar";
 export * from "./ui/sheet";
 export * from "./ui/tabs/tab-button";
 export * from "./ui/tabs/tabs";
